% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{compute_anomaly}
\alias{compute_anomaly}
\title{Compute anomaly from normals}
\usage{
compute_anomaly(monthly_data, normals_data)
}
\arguments{
\item{monthly_data}{}

\item{normals_data}{}
}
\value{
Data frame with anomaly
}
\description{
Compute anomaly from normals
}
