% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{define_color_scale}
\alias{define_color_scale}
\title{Define bins breaks and color scales to closely match NOAA's standard plots}
\usage{
define_color_scale(nclimgrid_data)
}
\arguments{
\item{nclimgrid_data}{Data frame with nClimGrid data, generated by \code{get_nclimgrid_monthly()} or \code{get_nclimgrid_normals()}}
}
\description{
Define bins breaks and color scales to closely match NOAA's standard plots
}
