% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_nclimgrid_monthly}
\alias{get_nclimgrid_monthly}
\title{Downloads and formats nClimGrid monthly data}
\usage{
get_nclimgrid_monthly(
  year,
  measurement = "tave",
  region = "us",
  wide = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{year}{Year, YYYY format}

\item{measurement}{Measurement type (i.e. "tave", "tmax", "tmin", "prcp")}

\item{region}{Region (i.e. "us", "ak")}

\item{wide}{Convert to wide format, default is FALSE}

\item{verbose}{Verbose, default is FALSE}
}
\value{
Data frame
}
\description{
Downloads and formats nClimGrid monthly data
}
