% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{month_to_name}
\alias{month_to_name}
\title{Labeller function to return month name for faceting plots}
\usage{
month_to_name(m)
}
\arguments{
\item{m}{Month number}
}
\value{
Month full name
}
\description{
Labeller function to return month name for faceting plots
}
