% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_measurement_data}
\alias{plot_measurement_data}
\title{Plot absolute measurement values}
\usage{
plot_measurement_data(
  nclim_data,
  subset_months = 1:12,
  show_states = TRUE,
  show_counties = FALSE,
  facet_cols = 2,
  title = NULL,
  subtitle = NULL,
  show_credit = TRUE
)
}
\arguments{
\item{subset_months}{List of numeric months to include, defaults to all (1:12)}

\item{show_states}{Show state outlines, default is TRUE}

\item{show_counties}{Show county outlines, default is FALSE}

\item{facet_cols}{Number of cols for faceting months, defaults to 2}

\item{title}{Optional plot title. A default plot title is generated based on the data type in nclimgrid_data}

\item{subtitle}{Optional plot title. A default plot title is generated based on the data type in nclimgrid_data}

\item{show_credit}{Show caption with dataset credit to NOAA, default is TRUE}

\item{nclimgrid_data}{Data frame with nClimGrid data, generated by get_nclimgrid_monthly() or get_nclimgrid_normals()}
}
\value{
Returns a ggplot object
}
\description{
Plot absolute measurement values
}
