% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_nclimgrid}
\alias{plot_nclimgrid}
\title{Plot absolute measurement values}
\usage{
plot_nclimgrid(
  nclimgrid_data,
  show_states = TRUE,
  show_counties = FALSE,
  facet_cols = 2,
  show_credit = TRUE
)
}
\arguments{
\item{nclimgrid_data}{Data frame with nClimGrid data, generated by get_nclimgrid_monthly() or get_nclimgrid_normals()}

\item{show_states}{Show state outlines, default is TRUE}

\item{show_counties}{Show county outlines, default is FALSE}

\item{facet_cols}{Number of cols for faceting months, defaults to 2}

\item{show_credit}{Show caption with dataset credit to NOAA, default is TRUE}
}
\value{
Returns a ggplot object
}
\description{
Plot absolute measurement values
}
