% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_nclimgrid_histogram}
\alias{plot_nclimgrid_histogram}
\title{Plot distribution}
\usage{
plot_nclimgrid_histogram(
  nclimgrid_data_a,
  nclimgrid_data_b = NULL,
  facet_cols = 3,
  show_credit = TRUE
)
}
\arguments{
\item{nclimgrid_data_a}{Data frame with nClimGrid data, generated by \code{get_nclimgrid_monthly()} or \code{get_nclimgrid_normals()}}

\item{nclimgrid_data_b}{(Optional) Data frame with nClimGrid data, generated by \code{get_nclimgrid_monthly()} or \code{get_nclimgrid_normals()}}

\item{facet_cols}{Number of cols for faceting months, defaults to 2}

\item{show_credit}{Show caption with dataset credit to NOAA, default is TRUE}
}
\description{
Plot distribution
}
