% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_title_builder}
\alias{plot_title_builder}
\title{Based on attributes of the nclimgrid_data data frame, generate a default title and subtitle for the plots.}
\usage{
plot_title_builder(nclimgrid_data)
}
\arguments{
\item{nclimgrid_data}{Data frame with nClimGrid data, generated by \code{get_nclimgrid_monthly()} or \code{get_nclimgrid_normals()}}
}
\value{
List with $title and $subtitle
}
\description{
The default title and subtitle can be overriden by adding \code{ggplot2::labs(title = "", subtitle = "")} to the plot object
}
