% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{progress_msg}
\alias{progress_msg}
\title{Print progress messages depending on verbose option}
\usage{
progress_msg(verbose, string)
}
\arguments{
\item{verbose}{}

\item{string}{Message to print if verbose is TRUE}
}
\description{
Print progress messages depending on verbose option
}
