% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_nclimgrid_archive_files}
\alias{read_nclimgrid_archive_files}
\title{Read the individual files inside a nClimGrid .tar.gz archive file}
\usage{
read_nclimgrid_archive_files(file_url)
}
\arguments{
\item{file_url}{URL to .tar.gz file}
}
\value{
List of data frames with file contents
}
\description{
Read the individual files inside a nClimGrid .tar.gz archive file
}
