% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standardize_units}
\alias{standardize_units}
\title{Standardizes the units from monthly nClimGrid datasets (C and mm) 
to the same units used by the normal historical averages (F and in)}
\usage{
standardize_units(values, measurement)
}
\arguments{
\item{values}{Vector of values}

\item{measurement}{Measurement type (i.e. "tmax", "tave", "tmin", "prcp)}
}
\value{
Values in standard units
}
\description{
Standardizes the units from monthly nClimGrid datasets (C and mm) 
to the same units used by the normal historical averages (F and in)
}
