% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_measurement}
\alias{validate_measurement}
\title{Validates if measurement names are valid
If no parameter is given, a list of valid options is returned}
\usage{
validate_measurement(measurement = NULL)
}
\arguments{
\item{measurement}{Measurement name}
}
\description{
Validates if measurement names are valid
If no parameter is given, a list of valid options is returned
}
