% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_normals_period}
\alias{validate_normals_period}
\title{Validates if normals period is a valid option
Normals exist in gridded format for some predefined periods}
\usage{
validate_normals_period(normals_period = NULL)
}
\arguments{
\item{normals_period}{Normals period range}
}
\description{
Validates if normals period is a valid option
Normals exist in gridded format for some predefined periods
}
