% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_region}
\alias{validate_region}
\title{Validates if region names are valid
If no parameter is given, a list of valid options is returned}
\usage{
validate_region(region = NULL)
}
\arguments{
\item{region}{Region name}
}
\description{
Validates if region names are valid
If no parameter is given, a list of valid options is returned
}
