% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_year}
\alias{validate_year}
\title{Validates if year is valid}
\usage{
validate_year(year)
}
\arguments{
\item{year}{Year, YYYY format}
}
\description{
nClimGrid stores the current and previous year only, other years are stored in a different format
and in another location - this script currently doesn't support fetching the archived years
}
