% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{define_color_scale}
\alias{define_color_scale}
\title{Define bins breaks and color scales to closely match NOAA's standard plots}
\usage{
define_color_scale(measurement, anomaly = FALSE)
}
\arguments{
\item{measurement}{Measurement type (i.e. tave, tmax, tmin, prcp)}

\item{anomaly}{TRUE for anomaly from normals color scale, FALSE for absolute measurement data}
}
\description{
Define bins breaks and color scales to closely match NOAA's standard plots
}
