% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_nclimgrid_histogram}
\alias{plot_nclimgrid_histogram}
\title{Plot distribution comparison}
\usage{
plot_nclimgrid_histogram(nclimgrid_data, ..., facet_cols = 2)
}
\description{
For a list of nClimGrid data frames, plot the distributions
}
