% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_title_builder}
\alias{plot_title_builder}
\title{Based on attributes of the nclim_data data frame, generate a default title and subtitle for the plots}
\usage{
plot_title_builder(nclimgrid_data)
}
\arguments{
\item{nclimgrid_data}{Data frame with nClimGrid data, generated by get_nclimgrid_monthly() or get_nclimgrid_normals()}
}
\value{
List with $title and $subtitle
}
\description{
Based on attributes of the nclim_data data frame, generate a default title and subtitle for the plots
}
